WIP: Split into more files. Allocate results in error.
